ENH: In order to make sure that there is only one CustomStatusProgressbar, implemented method getOrCreateCustomProgressBar: - all widgets that are using ModuleWidgetMixin as mixin class, have access to that method - for avoiding circular imports, created new widgets.py file where in future other widgets should be moved to as well
