+++ SlicerProstate/CMakeLists.txt
@@ -10,6 +10,7 @@ set(MODULE_PYTHON_SCRIPTS
   ${MODULE_NAME}Utils/constants.py
   ${MODULE_NAME}Utils/decorators.py
   ${MODULE_NAME}Utils/events.py
+  ${MODULE_NAME}Utils/widgets.py
   )
 
 set(MODULE_PYTHON_RESOURCES
