+++ SlicerProstate/SlicerProstateUtils/helpers.py
@@ -83,66 +83,6 @@ class SmartDICOMReceiver(ModuleLogicMixin, ParameterNodeObservationMixin):
     self.watchTimer.start()
 
 
-class CustomStatusProgressbar(qt.QWidget, ModuleWidgetMixin):
-
-  STYLE = "QWidget{background-color:#FFFFFF;}"
-
-  @property
-  def text(self):
-    return self.textLabel.text
-
-  @text.setter
-  def text(self, value):
-    self.textLabel.text = value
-
-  @property
-  def value(self):
-    return self.progress.value
-
-  @value.setter
-  def value(self, value):
-    self.progress.value = value
-
-  @property
-  def maximum(self):
-    return self.progress.maximum
-
-  @maximum.setter
-  def maximum(self, value):
-    self.progress.maximum = value
-
-  def __init__(self, parent=None):
-    qt.QWidget.__init__(self, parent)
-    self.setup()
-
-  def setup(self):
-    self.textLabel = qt.QLabel()
-    self.progress = qt.QProgressBar()
-    self.maximumHeight = slicer.util.mainWindow().statusBar().height
-    rowLayout = qt.QHBoxLayout()
-    self.setLayout(rowLayout)
-    rowLayout.addWidget(self.textLabel, 1)
-    rowLayout.addWidget(self.progress, 1)
-    self.setStyleSheet(self.STYLE)
-
-  def updateStatus(self, text, value=None):
-    self.text = text
-    if value is not None:
-      self.value = value
-
-  def reset(self):
-    self.text = ""
-    self.progress.reset()
-
-  def show(self):
-    slicer.util.mainWindow().statusBar().addWidget(self, 1)
-    qt.QWidget.show(self)
-
-  def hide(self):
-    slicer.util.mainWindow().statusBar().removeWidget(self)
-    qt.QWidget.hide(self)
-
-
 class SliceAnnotation(object):
 
   ALIGN_LEFT = "left"
